requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
anthropic>=0.40.0
google-api-python-client>=2.115.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
twilio>=9.0.0
sendgrid>=6.11.0
httpx>=0.26.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
